+++ database/migrations/2025_01_01_000002_create_reports_table.php
@@ -11,23 +11,25 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('reports', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('reporter_id')->nullable();
-            $table->unsignedBigInteger('reported_user_id');
-            $table->text('description');
-            $table->enum('status', ['pending', 'warning', 'resolved', 'close_account','dismissed'])->default('pending');
-            $table->timestamp('created_at')->useCurrent();
-            $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
-            
-            $table->foreign('reporter_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('reported_user_id')->references('id')->on('users')->onDelete('cascade');
-           
-            
+        if (!Schema::hasTable('reports')) {
+            Schema::create('reports', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('reporter_id')->nullable();
+                $table->unsignedBigInteger('reported_user_id');
+                $table->text('description');
+                $table->enum('status', ['pending', 'warning', 'resolved', 'close_account','dismissed'])->default('pending');
+                $table->timestamp('created_at')->useCurrent();
+                $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
+                
+                $table->foreign('reporter_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('reported_user_id')->references('id')->on('users')->onDelete('cascade');
+               
+                
 
-            $table->index(['status', 'created_at']);
-            $table->index(['reported_user_id', 'status']);
-        });
+                $table->index(['status', 'created_at']);
+                $table->index(['reported_user_id', 'status']);
+            });
+        }
     }
 
     /**
